feat: setup handlebars & set header to partial
So previously the header was repeated in each index, which was ok, but
now as I plan to add more common elements templating makes more sense so
I've setup handlebars. I've also converted the header part of each page
to a partial that accepts arguments, that decide if the header text
should be colored or not.