fix: footer positioning, overflow
Previous PR introduced a footer, but this footer would not be at the
bottom of the page on larger resolutions (zoom out). This commit fixes
that, by using a flex column layout on body along with min height 100vh
and growing child container.

To achieve this I had set body to a column layout, thereby breaking some
width styling in main container. Fixed this by using max-w instead of
width and the result is easier to understand, albeit still requires the
min-w-0 to be set so that containers shrink correctly.

The main element got wrapped in a parent div, so that we can retain the
flex direction we want. We want the child to grow in the horizontal but
the parent to grow in the vertical.