chore(deps): upgrade tailwindcss to v4.0
Ran the upgrade tool, fixed most issues, except for `w-100`, which isn't
supported anymore, replaced it with `w-full`.

One notable change is that the palette JSON isn't used anymore, the
upgrade tool figured out the parsing, and added the right colors, but
now we're not depending on upstream anymore. Maybe add some CSS vars to
the palette package.

ref: https://tailwindcss.com/docs/upgrade-guide